Move Seed Service into seed package
This commit also removes the ‘auto script’ logic and moves it into its own service named ‘SeedOnStartupService’. This is in preparation for changes to the seed on startup behaviour.
